+++ src/Services/Landing/Role/Result/EnableResult.php
@@ -1,0 +1,31 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Sally Fancen <vadimsallee@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Landing\Role\Result;
+
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Result\AbstractResult;
+
+class EnableResult extends AbstractResult
+{
+    /**
+     * Check if operation was successful
+     *
+     * @return bool
+     * @throws BaseException
+     */
+    public function isSuccess(): bool
+    {
+        return (bool)$this->getCoreResponse()->getResponseData()->getResult()[0];
+    }
+}